+++ src/dqx/api.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import functools
+import logging
 import threading
 import time
 import uuid
@@ -11,7 +12,7 @@ from typing import TYPE_CHECKING, Any, Protocol, cast, runtime_checkable
 import pyarrow as pa
 import sympy as sp
 
-from dqx import functions, get_logger
+from dqx import functions, setup_logger
 from dqx.analyzer import AnalysisReport, Analyzer
 from dqx.common import (
     AssertionResult,
@@ -38,7 +39,7 @@ CheckProducer = Callable[[MetricProvider, "Context"], None]
 CheckCreator = Callable[[CheckProducer], CheckProducer]
 
 
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 timer_registry = Registry()
 
 
@@ -323,6 +324,7 @@ class VerificationSuite:
         checks: Sequence[CheckProducer | DecoratedCheck],
         db: "MetricDB",
         name: str,
+        log_level: int = logging.INFO,
     ) -> None:
         """
         Initialize the verification suite.
@@ -335,6 +337,9 @@ class VerificationSuite:
         Raises:
             DQXError: If no checks provided or name is empty
         """
+        # Setting up the logger
+        setup_logger(level=log_level)
+
         if not checks:
             raise DQXError("At least one check must be provided")
         if not name.strip():
@@ -559,6 +564,7 @@ class VerificationSuite:
         Raises:
             DQXError: If no data sources provided or suite already executed
         """
+
         # Prevent multiple runs
         if self.is_evaluated:
             raise DQXError("Verification suite has already been executed. Create a new suite instance to run again.")