+++ tests/conftest.py
@@ -5,14 +5,14 @@ from unittest.mock import patch
 import pytest
 from rich.console import Console
 
-from dqx import get_logger
+from dqx import setup_logger
 from tests.fixtures.data_fixtures import commerce_data_c1, commerce_data_c2  # noqa: F401
 
 
 @pytest.fixture(autouse=True)
 def run_around_tests() -> Iterator:
     # Use dqx logger with default format
-    get_logger(level=logging.DEBUG, force_reconfigure=True)
+    setup_logger(level=logging.DEBUG, force_reconfigure=True)
     print("\n")
     yield